pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parakeet-qa"
version = "0.1.0"
description = "Answer questions from real-time audio using NVIDIA Parakeet TDT and Ollama"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]
dependencies = [
    "nemo_toolkit[asr]>=2.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "ollama>=0.4.0",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
parakeet-qa = "parakeet_qa.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
